DBZ-3566 Kafka test tooling abstracted and refactored into more extendable form
