+++ debezium-testing/debezium-testing-openshift/src/test/java/io/debezium/testing/openshift/mongodb/MongoConnectorIT.java
@@ -160,7 +160,7 @@ public class MongoConnectorIT extends ConnectorTestBase {
     @Order(8)
     public void shouldResumeStreamingAfterCrash() throws InterruptedException {
         operatorController.enable();
-        kafkaConnectController.waitForConnectCluster();
+        kafkaConnectController.waitForCluster();
         awaitAssert(() -> assertions.assertMinimalRecordsCount(dbServerName + ".inventory.customers", 7));
         awaitAssert(() -> assertions.assertRecordsContain(dbServerName + ".inventory.customers", "nibbles@test.com"));
     }
